PB/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
Pillow==9.3.0
requests==2.28.1
geopy==2.3.0
django-filter==22.1
django-cors-headers==3.13.0

asgiref==3.5.2
async-timeout==4.0.2
authentication==1.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
distlib==0.3.6
django-braces==1.0.0
django-extensions==1.1.1
filelock==3.8.0
Flask==2.2.2
geographiclib==2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
platformdirs==2.5.3
pycryptodome==3.15.0
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
virtualenv==20.16.6
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
